refactor: 전반적인 코드 수정
- 파일 경로 수정
- Image, a 태그를 Image, Link태그로 수정
