+++ components/common/modals/EditNameModal.tsx
@@ -1,5 +1,5 @@
 import { styled } from "styled-components";
-import closeIcon from "@/assets/icons/closeModal.png";
+import closeIcon from "@/public/assets/icons/closeModal.png";
 import Image from "next/image";
 import { BlueButton } from "../BlueButton";
 import { CommonModalProps } from "@/constants/commonTypes";
