+++ components/common/modals/SharedModal.tsx
@@ -1,5 +1,5 @@
 import styled from "styled-components";
-import closeIcon from "@/assets/icons/closeModal.png";
+import closeIcon from "@/public/assets/icons/closeModal.png";
 import Image from "next/image";
 import { CommonModalProps } from "@/constants/commonTypes";
 import { SOCIAL_ICONS } from "@/constants/socialIcon";
