+++ package.json
@@ -12,8 +12,9 @@
   "dependencies": {
     "next": "13.5.6",
     "react": "^18",
-    "react-dom": "^18",
-    "styled-components": "^6.1.8"
+    "react-dom": "^18.3.1",
+    "styled-components": "^6.1.8",
+    "swr": "^2.2.5"
   },
   "devDependencies": {
     "@svgr/webpack": "^8.1.0",
