+++ pages/Signup.tsx
@@ -2,9 +2,9 @@ import React, { useState, useEffect } from "react";
 import styled from "styled-components";
 import Image from "next/image";
 import Link from "next/link";
-import login_logo from "@/assets/login_logo.png";
-import icon_google from "@/assets/icons/icon_google.png";
-import icon_kakao from "@/assets/icons/icon_kakao.png";
+import login_logo from "@/public/assets/login_logo.png";
+import icon_google from "@/public/assets/icons/icon_google.png";
+import icon_kakao from "@/public/assets/icons/icon_kakao.png";
 import { BlueButton } from "@/components/common/BlueButton";
 import EmailPwdInput from "@/components/SignInUp/EmailPwdInput";
 import { postSignIn } from "@/api/api";
